1. quest can be joined when clicking on 'join' button. 2. User progress is shown on quest page if he has joined. 3.upload spot photo form has been added (does not send anything to server yet)

+++ src/main/java/ru/netcracker/registration/model/UserProgress.java
@@ -11,7 +11,7 @@ public class UserProgress {
     private Quest questByQuestId;
     private Date dateComplete;
     private Date takingDate;
-    private Integer userProgressId;
+    private Long userProgressId;
     private Collection<UserSpotProgress> userSpotProgressesByUserProgressId;
 
     @Basic
@@ -37,11 +37,11 @@ public class UserProgress {
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
     @Column(name = "user_progress_id")
-    public Integer getUserProgressId() {
+    public Long getUserProgressId() {
         return userProgressId;
     }
 
-    public void setUserProgressId(Integer userProgressId) {
+    public void setUserProgressId(Long userProgressId) {
         this.userProgressId = userProgressId;
     }
